requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Pillow==10.0.1
python-telegram-bot==20.5
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.6
